NodeDefsSelector/components: Refactor to separate files and fix imports
