+++ webapp/components/survey/NodeDefsSelector/AttributesSelector/AttributeSelector.scss
