+++ webapp/components/survey/NodeDefsSelector/AttributesSelector/AttributesSelector.scss
