+++ webapp/components/survey/NodeDefsSelector/AttributesSelector/index.js
@@ -1,0 +1,1 @@
+export { default } from './AttributesSelector'
