+++ webapp/components/survey/NodeDefsSelector/EntitySelector/EntitySelector.js
@@ -1,4 +1,4 @@
-import './entitySelector.scss'
+import './EntitySelector.scss'
 import React from 'react'
 import PropTypes from 'prop-types'
 import * as R from 'ramda'
