+++ webapp/components/survey/NodeDefsSelector/EntitySelector/EntitySelector.scss
