+++ webapp/components/survey/NodeDefsSelector/index.js
@@ -1,3 +1,5 @@
 export { default as NodeDefsSelectorView } from './NodeDefsSelector'
 
 export { default as NodeDefsSelectorAggregateView } from './NodeDefsSelectorAggregate'
+
+export { default as EntitySelector } from './EntitySelector'
