+++ webapp/loggedin/modules/analysis/step/view/components/entitySelector.js
@@ -1,4 +1,4 @@
-import './entitySelector.scss'
+import './EntitySelector.scss'
 import React from 'react'
 import PropTypes from 'prop-types'
 import { useDispatch } from 'react-redux'
